app/requirements.txt
customtkinter
PyYAML
pyperclip
Pillow
google-genai
